refactor(models): split monolithic __init__.py into modular structure
リファクタリング概要:
- 1,308行の単一ファイルを10個の適切なサイズのモジュールに分割
- 単一責任の原則に基づいた明確な責務分離
- 保守性、可読性、拡張性を大幅に向上

ファイル構成:
app/models/
├── __init__.py          (95行)   - 公開API定義
├── types.py             (75行)   - 型定義 (既存)
├── errors.py            (13行)   - エラークラス
├── database.py          (62行)   - DB接続・セッション管理
├── base.py              (93行)   - SQLAlchemy基底クラス
├── stock_data.py        (368行)  - 株価データモデル (8モデル)
├── master.py            (196行)  - マスタデータモデル (2モデル)
├── batch.py             (219行)  - バッチ実行モデル (2モデル)
└── crud/
    ├── __init__.py      (7行)
    └── stock_daily.py   (432行)  - StockDaily CRUD操作

主な変更点:
- エラークラスを errors.py に分離
- DB設定とセッション管理を database.py に分離
- SQLAlchemyベースクラスを base.py に分離
- モデル定義を責務ごとに分離 (stock_data, master, batch)
- CRUD操作を crud/ ディレクトリに分離

テスト修正:
- モックパスを app.models.SessionLocal から
  app.models.database.SessionLocal に変更
- tests/unit/models/test_models.py (2箇所)
- tests/unit/utils/test_database_connection.py (3箇所)

検証結果:
- 全テスト成功: 87/87 (100%)
- Flake8: エラーなし
- 後方互換性: 完全に維持
