+++ app/models/crud/__init__.py
@@ -1,0 +1,8 @@
+"""CRUD操作モジュール."""
+
+from app.models.crud.stock_daily import StockDailyCRUD
+
+
+__all__ = [
+    "StockDailyCRUD",
+]
